+++ src/world/astronomy/orbit.rs
@@ -2,7 +2,7 @@ use std::f32::consts::{PI, TAU};
 
 use bevy::prelude::*;
 
-use crate::core::resources::{game_time::GameTime, state::GameState};
+use crate::core::resources::game_time::GameTime;
 
 pub const GRAVITATIONAL_CONSTANT: f32 = 6.67430e-11; // https://en.wikipedia.org/wiki/Gravitational_constant#Modern_value
 pub const ORBITAL_PERIOD_SCALING_FACTOR: f32 = 1.0e-13;
@@ -46,17 +46,6 @@ impl Mass {
 impl Default for Mass {
     fn default() -> Self {
         Self::ZERO
-    }
-}
-
-pub struct OrbitPlugin;
-impl Plugin for OrbitPlugin {
-    fn build(&self, app: &mut App) {
-        app.add_systems(
-            Update,
-            (orbitable_update_system, orbital_positioning_system)
-                .run_if(in_state(GameState::Active)),
-        );
     }
 }