+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Migrated core and temp systems to root systems.rs file for better overview
+- Migrated core, temp, ui and world systems to root systems.rs file for better overview
 
 ## [0.0.19] - 2023-10-31